+++ zenovis/xinxinoptix/optixTriMesh.h
@@ -30,7 +30,7 @@ struct MeshObject {
 
     std::vector<float3> vertices{};
     std::vector<uint3>  indices {};
-    std::vector<uint>   mat_idx {};
+    std::vector<uint16_t> mat_idx{};
 
     std::vector<float2> g_uv;
     std::vector<ushort3> g_clr, g_nrm, g_tan;
@@ -92,7 +92,7 @@ using raii = xinxinoptix::raii<T>;
         return std::vector { d_idx.handle, d_uv.handle, d_clr.handle, d_nrm.handle, d_tan.handle };
     }
 
-    void buildGas(OptixDeviceContext context, const std::map<std::string, uint>& mtlidlut) {
+    void buildGas(OptixDeviceContext context, const std::map<std::string, uint16_t>& mtlidlut) {
 
         auto buffers = this->aux();
         std::reverse(buffers.begin(), buffers.end());
